+++ src/base/parameter.jl
@@ -64,7 +64,7 @@ function get_parameter(
     return @load_preference("$(e)", default)
 end
 # Handle empty defaults
-get_parameter(::Symbol, ::Val{:default}) = nothing
+get_parameter(::Val, ::Val{:default}) = nothing
 get_parameter(::Val{:Mode}, ::Val{:default}) = nothing
 get_parameter(::Val{:KeywordsErrorMode}, ::Val{:default}) = "warn"
 """